refactor(v0.1.0): split cli parsing into separate classes
